+++ Assets/MonoBehaviours/Views/GeometryEditorViews/GeometryEditorUI.cs
@@ -5,44 +5,10 @@ public class GeometryEditorUI: MonoBehaviour {
 
     public GeometryEditMode controller;
 
-    public GameObject graphicsPropertiesView;
-
-    public GameObject activeGraphicsPropertiesView = null;
-
     void Start() {
 
         controller.view = this;
 
-
-    }
-
-    void Update() {
-
-        if (Input.GetKeyDown(KeyCode.Escape)) {
-
-            if (activeGraphicsPropertiesView != null) {
-                CloseGraphicsPropertyView();
-            }
-
-        }
-
-    }
-
-    public void OnClickGraphicsProperitesButton() {
-
-        if (controller.selectedTiles.Count == 0) { return; }
-
-        if (activeGraphicsPropertiesView != null) {
-            CloseGraphicsPropertyView();
-        } else {
-
-            activeGraphicsPropertiesView = Instantiate(graphicsPropertiesView);
-
-            activeGraphicsPropertiesView.GetComponent<GraphicsPropertiesView>().controller = controller;
-
-            activeGraphicsPropertiesView.transform.SetParent(transform.parent, false);
-
-        }
 
     }
 
@@ -58,16 +24,6 @@ public class GeometryEditorUI: MonoBehaviour {
 
     }
 
-    void CloseGraphicsPropertyView() {
-        Destroy(activeGraphicsPropertiesView);
-        controller.selectedSection.RefreshMesh();
-    }
-
-    public void OnClickDuplicateTileGraphicsButton() {
-
-        controller.DuplicateTileGraphics();
-
-    }
 
 }
