--- Assets/MonoBehaviours/Views/GraghicsPropertiesViews/GraphicsPropertiesView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3984698e82a2ce646bf03915d5fd3ce3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
