--- Assets/MonoBehaviours/Views/GraghicsPropertiesViews/TextureOffsetItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 048e766903b6d1a45adfe3cbe9d74f3a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
