--- Assets/MonoBehaviours/Views/TextureEditView.cs
@@ -1,9 +1,0 @@
-﻿
-
-using UnityEngine;
-
-public class TextureEditView: MonoBehaviour {
-
-    public TextureEditMode controller;
-
-}
