+++ Assets/MonoBehaviours/Views/TextureEditorViews/TextureCoordinatePoint.cs.meta
