+++ Assets/MonoBehaviours/Views/TextureEditorViews/TextureUVMapper.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 324c00dd533d41044b1d209ddbe98cd5
+guid: 78aeae5b3b8c13949a3d020bdee8ea7c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
