+++ Assets/MonoBehaviours/Views/TextureEditorViews/TileTexturePreview.cs.meta
