--- Assets/Views/Graphics Preset Item.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9c34f505efb78cf43b1dff87d0de015a
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
