--- Assets/Views/Texture Offset Item.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 354420d54a4fb5244b6c839f3d4ac41e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
